+++ malashenko.dmitrii/S1/functions.hpp
@@ -11,15 +11,15 @@ namespace malashenko
   template< class T > class LCIter;
   template< class T > class List;
 
-  using pair_t = std::pair< std::string, List< int > >;
+  using pair_t = std::pair< std::string, List< size_t > >;
 
   void getData(std::istream& in, List< pair_t >& res);
   size_t getMaxSeqSize(const List< pair_t >& list);
   void printSeqNames(std::ostream& out, const List< pair_t >& list);
-  void getTransedSeq(List< List< int > >& res, const List< pair_t >& list);
-  void printNewSeqs(std::ostream& out, const List< List< int > >& list);
-  void countSums(const List< List< int > >& list, List< int >& sums);
-  void printSums(std::ostream& out, const List< int >& sums);
+  void getTransedSeq(List< List< size_t > >& res, const List< pair_t >& list);
+  void printNewSeqs(std::ostream& out, const List< List< size_t > >& list);
+  void countSums(const List< List< size_t > >& list, List< size_t >& sums);
+  void printSums(std::ostream& out, const List< size_t >& sums);
 }
 
 #endif
